Correct bitrate adjustment
Until now the script adjusted the video bitrate like this
    v = v_old * target_size/file_size
This calculation is correct when there's only a video stream. Once you add audio though, the file's size becomes the product of both the audio and the video bitrate. So any adjustment to the overall bitrate, based on the output file over- or undershooting the target file size, applies to both the audio and video bitrate. As the audio bitrate is constant (for a single file), the audio bitrate adjustment needs to be added to or subtracted from the video bitrate as well. This isn't done right now, which leads to a certain error. The video bitrate doesn't get reduced or raised as much as it should be. This isn't much of a problem as long as the audio bitrate is small compared to the video bitrate (which is normal), but it becomes noticeable for low target sizes.

The new calculation
    v = v_old * target_size/file_size - (target_size/file_size - 1) * a
compensates for the missing audio bitrate change by adjusting the video bitrate further.