Store ANSI escape codes in own class
The whole color management is a mess. This is the first step to clean
things up by switching from a dictionary to a class. Its disable()
function allows us to get rid of all those "if opts.color" statements.

The way color is applied by separate print calls, as well as the lack
of native support for ANSI escape codes on Windows is still problematic,
but will be fixed in later commits.