Overhaul multiline help/info texts
Triple-quoted strings are great for big help or info dumps, but you
can't indent them, as all inserted whitespaces would be printed as well.
The textwrap module helps with that by providing a way to strip all
common whitespaces, allowing us to align these strings with the rest of
the code.

In addition to introducing textrwap, the commit also removes the smaller
info text functions. Since msg() handles the verbosity level, there's no
benefit in putting those texts in their own functions (they are all only
a few lines long and at most get called twice within the script).
print_options() is the exception. That behemoth deserves its own place.

The drawback with the smaller info texts is that they are supposed to be
indented with 2 whitespaces on each line. These would also be stripped.
To fix this we use textwrap's indent function, right after using its
dedent function. Sounds a bit dumb, but as long as it works...