Simplify Values.calc_path()
This was achieved in several steps:

  1) Don't resolve the output directory.

There's no need for the object to hold any information about the output
directory. We only need it in resolve_path(), so let it deal with
determining the output directory instead.

  2) Check for reserved filenames in check_options().

Doing the check in calc_path() was always somewhat unfortunate, because
it triggers a hard exit, but the function only gets called once the
script goes to the next input file. So the script could potentially quit
in the middle of a session. Now input_list gets checked right after the
command line was parsed, exiting the script early on in case of reserved
filenames.

  3) f-strings

Gotta love them.