Refactor Values and Settings classes
RestrictedWebM always had a central problem (one that's still present,
even with the now fourth iteration of the script... or was it the
fifth?). There are so many properties and settings to keep track of. All
of which influence other properties and settings throughout the script
or change with each iteration.

I want a clean way to handle information and the best way in my opinion
is to use a general file class. One that has access to all important
properties, but is also able to assemble and handle the necessary
FFmpeg options.
That's the job of the ConvertibleFile class (new version of Settings).
Unfortunately it would quickly turn into a behemoth, if I were to use it
to store all necessary properties and methods by itself. Instead we use
another class FileInfo (new version of Values). It's purpose is to
gather and store all important input and output properties, as well as
updating properties which change between encoding attempts.
The important difference to the prior Values/Settings approach is, that
ConvertibleFile holds a FileInfo object and has access to all
properties.

As part of the refactor many functions were rewritten or at least
changed. Still, there are many planned improvements yet missing. Most
notably the assembling of the final command should be done by the
ConvertibleFile object and not call_ffmpeg().

These changes are yet barely tested.