Overhaul terminal output
Designing good looking terminal output isn't as easy as it looks. Or at
least it isn't for someone like me, who has absolutely no eye for these
things.

This commit tries to fix and tweak various aspects of the terminal
output to make it less confusing, easily searchable (scrollback) and
more pleasing to look at.

  -) Less Colors

The most obvious one. To say that the default output looked like a clown
threw up on the screen is probably an understatement. I think colors are
beneficial to loosen up big walls of information and to make the most
important aspects stand out, but that effect is lost when pretty much
every line is a different color.
So stop colorizing every bit of information and focus on the important
bits, like values that change between encoding attempts, the input
filename and in verbose mode the section headers and best file size.

  -) Bold text

The script uses a level system to classify the importance of a message.

  2 -> only shown in verbose mode (FFmpeg settings, size info, ...)
  1 -> shown in default mode (filename, attempt info)
  0 -> even shown in quite mode (errors and warnings)

The lower the level of a message the more important it is. Errors and
warnings are always in color, so they already stand out in the now much
less colorful output. Level 1 information, however, may blur in with
level 2 information in verbose mode. To avoid this, level 1 information
is bold in verbose mode and normal during default mode.

  -) Readable size information

Let's be honest, there's no benefit to printing the size information in
Bytes, except that it allowed me to be lazy by not having to format the
values. Printing them in MB gives a far better understanding of how well
the script progresses.
In addition to the formatted values, the best achieved size is now
printed with a variable color. Red signals too large, yellow too small
and green in range.

  -) File progress

Printing the filename in a unique color gives a nice overview where a
new file conversion starts, but there was no clear indication of the
overall progress for larger batches. Remedy that by introducing a
counter before the filename.

  -) FFmpeg progress during quiet mode

FFmpeg verbosity can be changed with its own option and is therefore
independent from the general verbosity level. This is fine in general,
but if the user already specified quiet script execution, chances are
they don't want FFmpeg output either. So -q/--quiet now also changes
the FFmpeg verbosity to "warning".
This also solves another problem. FFmpeg output behaves rather weird, if
colorized output is used right beforehand. Even if the stream's color
was reset, FFmpeg will use the previous color, unless a newline is
printed before it. That's not a problem during default or verbose mode,
but the quiet mode only shows errors and warnings, which all come in
either yellow or red. So once a warning or error appears, all FFmpeg
output will inherit its color.

  -) No exclamation mark for warnings (except user interrupt)
  -) Absolute output path in verbose mode
  -) Stand-in phrase for FFmpeg conversion, if FFmpeg output is disabled
     during default/verbose mode
  -) Print size warnings in one line