Switch to argparse for command-line parsing
I do not like argparse. In fact, you could say I hate it. It works well
enough if all you do is assign values to variables, but as soon as you
want to do something more advanced (e.g. slightly customized help
messages or - god forbid - custom actions) it gets tedious to work with.
But I also have to be realistic here. Manual parsing just doesn't scale
well, especially with a language that doesn't provide case statements.
Of course the always looming threat of a forgotten value check in
check_options() wasn't great either.

So I bit the bullet and switched to argparse.

The original Options class (now with a different name) is still present,
mainly because I want all defaults at the very beginning of the script,
so that they are readily available. The advanced option defaults are now
in parse_cli() though.
Unfortunately I still have to do some additional option checking, since
some pitfalls can only be detected after all options are set. Sometimes
it's also not worth it to introduce a new type function, so that
argparse can check on assignment, when it only applies to one option.
Wouldn't it be amazing, if you could just pass argparse a condition,
instead of having to define a new type?

On a side note, this commit also removes the --no-color flag, as it
didn't disable colorized output for default error messages and would
leave your terminal output colorized in case of such a message.