Imply subtitle mapping with --burn-subs
--burn-subs was used to discard subtitle output with -sn, but it didn't
map the subtitles in the first place, which is necessary, if one wants
to use a filter for hardsubbing. So instead of just using --burn-subs,
one had to use --subtitles --burn-subs (the former to map, the latter to
discard). That's tedious and simply unnecessary. Now --burn-subs also
maps input subtitles.

I could swear this was already the case at one point.