Remove repeated calls of out_image_subs()
In general it's best to keep the number of FFmpeg calls (even if it's
a quick one, which only tries to copy 1 second of subtitles) to a
minimum.
I also think it's more straightforward to store the state of output
image subtitles in a separate variable, instead of checking against the
extension.