Reintroduce stream order check
The previous test was commented out for a while, since the ffprobe info
JSON (required to get the stream types) is now confined to the
initialization of FileInfo objects.

Testing if the 1st stream is a video needs to be done in FileInfo
directly, because other methods would fail if that assumption is not
true. The other tests could be done at a later point, but I don't want
to separate them, since they are all input related.
Unfortunately the validity needs to be checked in 3 places. In the init
for FileInfo (would fail because of wrong info), in the init for
ConvertibleFile (would fail because of missing info) and the main file
loop to skip the current file. The check in ConvertibleFile could be
removed, but I don't want the FileInfo object being directly accessible
in main() and with the current approach it would only require minimal
changes to introduce a separate validity check to ConvertibleFile.