Change to absolute output paths
Previously the script changed into each output directory and used
relative paths for the final output and any temporary files. Now it uses
absolute output paths instead.
This allows the usage of FFmpeg/ffprobe executables in the same location
as the script, without having to add them to your PATH, when using CMD.
Un*x users as well as Windows users running Powershell still have to add
the executables to their PATH. The only difference for them is that the
2 pass FFmpeg log will now appear in the same location as the script (it
will still be cleaned up though).

Fixes: #2